fix(pr): resolve review and CI issues for #876
- Add int overload for IsConstValue to avoid int64_t/double ambiguity
- Fix misc-include-cleaner: add direct includes in memref_collectors.h,
  remove unused includes in insert_sync_pass.cpp and
  interchange_chunk_loops_pass.cpp, add transform_utils.h include in
  outline_incore_scopes_pass.cpp

+++ src/ir/transforms/op_conversion_registry.cpp
@@ -50,7 +50,7 @@ ExprPtr MakeShapesTuple(const std::vector<ExprPtr>& shape, const Span& span) {
   return std::make_shared<MakeTuple>(shape, span);
 }
 
-bool IsConstOne(const ExprPtr& expr) { return IsConstValue(expr, static_cast<int64_t>(1)); }
+bool IsConstOne(const ExprPtr& expr) { return IsConstValue(expr, 1); }
 
 // Detect row-broadcast pattern: [M, N] op [M, 1] or [M, 1] op [M, N]
 // Returns {wider_arg_idx, narrower_arg_idx} if broadcast detected, empty otherwise
